# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LiftGO API",
    "version": "1.0.0",
    "description": "Generated from Zod schemas. Do not edit manually."
  },
  "paths": {
...
    "/api/v1/devices/register": {
      "post": {
        "tags": [
          "Devices"
        ],
        "summary": "Register push device token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "platform": {
                    "type": "string",
                    "enum": [
                      "ios",
                      "android",
                      "web"
                    ]
                  },
                  "appVersion": {
                    "type": "string"
                  },
                  "deviceName": {
                    "type": "string"
                  }
                },
                "required": [
                  "token",
                  "platform"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device registered"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}